Implementado o método lista top3

+++ src/main/java/br/com/everaldoboscatto/LiterAlura/model/DadosLivro.java
@@ -4,9 +4,10 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 import java.util.List;
 
+// Recebe os dados vindos da API
 @JsonIgnoreProperties(ignoreUnknown = true)
 public record DadosLivro(@JsonAlias({"title"}) String titulo,
                          @JsonAlias({"authors"}) List<DadosAutor> autores,
                          @JsonAlias("languages") List<String> idiomas,
-                         @JsonAlias("download_count") Integer numeroDownloads) {
+                         @JsonAlias("download_count") Double numeroDownloads) {
 }
